+++ js/main.js
@@ -1,4 +1,4 @@
 import {changeScreen} from './module-mangment-dom.js';
-import mainScreen from './module-main-screen.js';
+import mainScreen from './screens/module-main-screen.js';
 
 changeScreen(mainScreen());
