--- js/module-game-3-screen.js
@@ -1,61 +1,0 @@
-import {renderTemplate} from './module-mangment-dom.js';
-import {controlGameScreens, recordUserAnswer} from './game.js';
-import dataGame from './data-game.js';
-
-const CORRECT_ANSWER = `paint`;
-
-/** =========================================
- * возврашает шаблон с данными
- * @param {Array} arrImages
- * @param {String} statsAnswersStr
- * @return {String}
- */
-const template = (arrImages, statsAnswersStr) => {
-  let htmlImages = ``;
-  arrImages.forEach((item, index) => {
-    htmlImages += `<div class="game__option">
-              <img src="${item.src}" data-type="${item.imageType}" alt="Option ${index}" width="304" height="455">
-            </div>`;
-  });
-  return `
-    <section class="game">
-      <p class="game__task">Найдите рисунок среди изображений</p>
-      <form class="game__content game__content--triple">
-        ${htmlImages}
-      </form>
-      <ul class="stats">
-        ${statsAnswersStr}
-      </ul>
-    </section>
-  `;
-};
-/** при выборе ответа в форме, переключение экрана
- * @param {Event} evt
- * @param {Object} state
- */
-const clickFormHandler = (evt, state) => {
-  const target = evt.target;
-  const selectUserAnswer = target.getAttribute(`data-type`);
-
-  const newState = recordUserAnswer(CORRECT_ANSWER === selectUserAnswer, state);
-  controlGameScreens(newState, dataGame);
-};
-/** =========================================
- * экспорт
- * @param {Object} state
- * @param {Array} arrImages
- * @param {String} statsAnswersStr
- * @return {HTMLElement} element
- */
-export default (state, arrImages, statsAnswersStr) => {
-  const element = renderTemplate(template(arrImages, statsAnswersStr));
-  const imgs = element.querySelectorAll(`.game__content img`);
-
-  imgs.forEach((item) => {
-    item.addEventListener(`click`, (evt) => {
-      clickFormHandler(evt, state);
-    });
-  });
-
-  return element;
-};
