--- js/module-main-screen.js
@@ -1,28 +1,0 @@
-import {renderTemplate, changeScreen} from './module-mangment-dom.js';
-import welcomeScreen from './module-welcome-screen.js';
-
-const MAIN_SCREEN = `
-  <section class="intro">
-    <button class="intro__asterisk asterisk" type="button"><span class="visually-hidden">Продолжить</span>*</button>
-    <p class="intro__motto"><sup>*</sup> Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf Sparnaay.</p>
-  </section>
-`;
-
-/** изменение sreen по клику и css анимация исчезновения/появления
-*
-*/
-const clickBtnHandler = () => {
-  changeScreen(welcomeScreen());
-};
-/** =========================================
-* экспорт
-* @return {HTMLElement} element
-*/
-export default () => {
-  const element = renderTemplate(MAIN_SCREEN);
-  const btnIntroAsterisk = element.querySelector(`.intro__asterisk`);
-
-  btnIntroAsterisk.addEventListener(`click`, clickBtnHandler);
-
-  return element;
-};
