--- js/module-stats.js
@@ -1,13 +1,0 @@
-import {renderTemplate} from './module-mangment-dom.js';
-
-const statsTemplate = (listItems) => `<ul class="stats">
-        ${listItems}
-      </ul>
-    </section>
-`;
-
-export default (listItems) => {
-  const element = renderTemplate(statsTemplate(listItems));
-
-  return element;
-};
